Seperating functionality into discrete files
